Make loans fetch from database

+++ lib/data_source/models/loan_model.dart
@@ -1,7 +1,7 @@
 class LoanModel {
   String? loanNumber;
-  double? principalAmount;
-  double? loanFees;
+  int? principalAmount;  // should be double but api returns int.
+  int? loanFees;  // should be double but api returns int.
   int? clientId;
   String? clientName;
   int? loanTypeId;
